+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add `TransactionRetry.before_retry` configuration option to run Proc before transaction retry
 * Add `TransactionRetry.retry_on` configuration option to include more Errors to retry on
 * Update dependency activerecord >= 5.1
 * Update dependency ruby 2.2.2